Create Day 15: Linked List.java
You are given a class Node in the editor which has one instance pointer next pointing to next node and an integer data to store the data in Node.

You are also given a pointer head pointing to the head node of a linked list and an integer data to add to the list. Create a new node with the given integer. Insert this node at the tail of the linked list and return the head node. The given head pointer may be null, meaning that the initial list is empty.

Code for input/output is already handled in the editor. You have to complete the function insert given in the editor. It takes two arguments: the head node of the linked list and an integer data to be inserted.