Create  Day 14: All about Scope.java
In this challenge, you will create a program that takes N non-negative integers as input and finds the greatest absolute difference between two of the N integers, and then print this difference to the console.

There is a class Difference given in the editor with one private instance array elements which stores the N non negative integers and public integer maxDifference to store the greatest absolute difference between the two of the N integers.

Code for handling input/output is already given in the editor. Your task is to write the constructor for the class Difference and the method computeDifference which finds the greatest absolute difference between any two numbers in the input array and stores it in maxDifference.