Create Day 13: Abstract Classes.java
An abstract class is a class that cannot be instantiated, meaning you cannot create a new instance of it. Attempting to instantiate an abstract class (such as abstract class Book, provided for you in the editor) like so:

Book new_novel=new Book("The Alchemist","Paulo Coelho"); 

results in error: Book is abstract; cannot be instantiated. This type of class is only meant to serve as a base or blueprint for connecting the subclasses that inherit it.

To use an abstract class, you must create its subclass and then instantiate the subclass. Any abstract methods declared in the abstract class (such as abstract void display() in abstract class Book) must be implemented by the subclass.

In the editor we have provided the abstract Book class and a Solution class. In the Solution class we created instance of a class called MyBook. Your task is to write just the MyBook class. The class MyBook mustn't be public.