Create Day 19: Interfaces.java
Here you are given an interface AdvancedArithmetic which contains a method signature
int divisorSum(int n). (The divisorSum function just takes an integer as input and return the sum of all its divisors.) Your only task is to write a class Calculator which implements the interface.

Note : The class Calculator shouldn't be public.