Create Day 22: Binary Search Trees.java
The height of a binary tree is the number of nodes on the largest path from root to any leaf. You are given a pointer root pointing to the root of a binary search tree. Return the height of the tree.
You only have to complete the function getHeight given in the editor.